ai-trading/mcp/pyproject.toml
[project]
name = "hummingbot_mcp"
version = "0.1.0"
description = "MCP server for Hummingbot API integration - manage crypto trading across multiple exchanges"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "hummingbot-api-client==1.1.4",
    "fastmcp>=2.11.0",
    "pydantic>=2.11.2",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/hummingbot/mcp"
Repository = "https://github.com/hummingbot/mcp"
Issues = "https://github.com/hummingbot/mcp/issues"

[project.scripts]
mcp-hummingbot = "hummingbot_mcp:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["hummingbot_mcp"]

[tool.ruff]
line-length = 140

[tool.ruff.lint]
select = [
    "E",      # https://docs.astral.sh/ruff/rules/#error-e
    "F",      # https://docs.astral.sh/ruff/rules/#pyflakes-f
    "I",      # https://docs.astral.sh/ruff/rules/#isort-i
    "FA",     # https://docs.astral.sh/ruff/rules/#flake8-future-annotations-fa
    "UP",     # https://docs.astral.sh/ruff/rules/#pyupgrade-up
    "RUF100", # https://docs.astral.sh/ruff/rules/#ruff-specific-rules-ruf
]
ignore = ["UP031"] # https://docs.astral.sh/ruff/rules/printf-string-formatting/

[tool.ruff.lint.isort]
combine-as-imports = true

[dependency-groups]
dev = ["ruff"]
